+++ tests/testthat/test_aaa.R
@@ -1,2 +1,0 @@
-# Prevent warning from being raised by loggit()'s first call of the session
-.config$seenmessage <- TRUE