+++ tests/testthat/test_utils.R
@@ -1,5 +1,6 @@
 context("Utility functions")
 
+
 test_that("bind_rows_loggit works", {
   df1 <- data.frame(a = c(1, 2, 3))
   df2 <- data.frame(b = c(1, 2, 3))
@@ -9,3 +10,11 @@ test_that("bind_rows_loggit works", {
   )
   expect_equal(bind_rows_loggit(df1, df2), df_out)
 })
+
+
+test_that("log_rotate works", {
+  fail()
+})
+
+
+cleanup()